+++ src/Ai/Base/ActionContext.h
@@ -29,7 +29,6 @@
 #include "EmoteAction.h"
 #include "FishingAction.h"
 #include "FollowActions.h"
-#include "GenericActions.h"
 #include "GenericSpellActions.h"
 #include "GiveItemAction.h"
 #include "GreetAction.h"
@@ -48,6 +47,7 @@
 #include "NewRpgOutdoorPvP.h"
 #include "NonCombatActions.h"
 #include "OutfitAction.h"
+#include "PetsAction.h"
 #include "PositionAction.h"
 #include "PullActions.h"
 #include "RandomBotUpdateAction.h"
