+++ src/Ai/Base/Actions/PetsAction.h
@@ -23,4 +23,26 @@ private:
     std::string defaultCmd;
 };
 
+class TogglePetSpellAutoCastAction : public Action
+{
+public:
+    TogglePetSpellAutoCastAction(PlayerbotAI* ai) : Action(ai, "toggle pet spell") {}
+    virtual bool Execute(Event event) override;
+};
+
+class PetAttackAction : public Action
+{
+public:
+    PetAttackAction(PlayerbotAI* ai) : Action(ai, "pet attack") {}
+    virtual bool Execute(Event event) override;
+};
+
+class SetPetStanceAction : public Action
+{
+public:
+    SetPetStanceAction(PlayerbotAI* botAI) : Action(botAI, "set pet stance") {}
+
+    bool Execute(Event event) override;
+};
+
 #endif
