+++ src/Ai/Raid/BWL/BWLMultipliers.cpp
@@ -8,7 +8,6 @@
 #include "BWLActions.h"
 #include "BWLHelpers.h"
 #include "ChooseTargetActions.h"
-#include "GenericActions.h"
 #include "HunterActions.h"
 #include "Playerbots.h"
 #include "ReachTargetActions.h"
