+++ src/Ai/Raid/EoE/EoEMultipliers.cpp
@@ -12,7 +12,6 @@
 #include "EoEActions.h"
 #include "EoETriggers.h"
 #include "FollowActions.h"
-#include "GenericActions.h"
 #include "GenericSpellActions.h"
 #include "MovementActions.h"
 #include "PaladinActions.h"
