+++ src/Ai/Raid/ICC/Action/ICCActions_GSB.cpp
@@ -5,7 +5,6 @@
  */
 
 #include "EquipAction.h"
-#include "GenericActions.h"
 #include "GenericSpellActions.h"
 #include "ICCActions.h"
 #include "ICCTriggers.h"
