+++ src/Ai/Raid/ICC/ICCMultipliers.cpp
@@ -11,7 +11,6 @@
 #include "DruidActions.h"
 #include "DruidBearActions.h"
 #include "FollowActions.h"
-#include "GenericActions.h"
 #include "GenericSpellActions.h"
 #include "HunterActions.h"
 #include "ICCActions.h"
@@ -20,6 +19,7 @@
 #include "MageActions.h"
 #include "MovementActions.h"
 #include "PaladinActions.h"
+#include "PetsAction.h"
 #include "PlayerbotAI.h"
 #include "PriestActions.h"
 #include "ReachTargetActions.h"
