+++ src/Ai/Raid/Kara/KaraMultipliers.cpp
@@ -10,12 +10,12 @@
 #include "DKActions.h"
 #include "DruidActions.h"
 #include "FollowActions.h"
-#include "GenericActions.h"
 #include "HunterActions.h"
 #include "KaraActions.h"
 #include "KaraHelpers.h"
 #include "MageActions.h"
 #include "PaladinActions.h"
+#include "PetsAction.h"
 #include "Playerbots.h"
 #include "PriestActions.h"
 #include "ReachTargetActions.h"
