+++ src/Ai/Raid/Naxx/Action/NaxxActions.h
@@ -9,7 +9,6 @@
 
 #include "Action.h"
 #include "AttackAction.h"
-#include "GenericActions.h"
 #include "MovementActions.h"
 #include "NaxxBossHelper.h"
 #include "PlayerbotAI.h"
