+++ src/Ai/Raid/Naxx/NaxxMultipliers.cpp
@@ -10,7 +10,6 @@
 #include "DruidActions.h"
 #include "DruidBearActions.h"
 #include "FollowActions.h"
-#include "GenericActions.h"
 #include "GenericSpellActions.h"
 #include "HunterActions.h"
 #include "MageActions.h"
@@ -18,6 +17,7 @@
 #include "NaxxActions.h"
 #include "NaxxSpellIds.h"
 #include "PaladinActions.h"
+#include "PetsAction.h"
 #include "PriestActions.h"
 #include "ReachTargetActions.h"
 #include "RogueActions.h"
